% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{estimate_faprob}
\alias{estimate_faprob}
\title{Estimate false alarm probability}
\usage{
estimate_faprob(N, change_type, t, k, tau, m = 200)
}
\arguments{
\item{N}{window sizes; vector of integers}

\item{change_type}{decides the type of algortihm for the corresponding
window sizes; vector of "jump" and "kink"}

\item{t}{threshold; numerical vector}

\item{k}{number of historical data available; integer}

\item{tau}{estimates the probaility of a false alarm before tau; integer}

\item{m}{number of repetitions; integer}
}
\value{
estimated false alarm probability
}
\description{
Estimates the false alarm probability of the Sequential Jump and Kink Detector,
that is the probability of a false alarm befor tau
}
