% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{fa}
\alias{fa}
\title{False alarm}
\usage{
fa(N, change_type, t, k, tau)
}
\arguments{
\item{N}{window sizes; vector of integers}

\item{change_type}{decides the type of algorithm for the corresponding
window sizes; vector of "jump" and "kink"}

\item{t}{threshold; numerical vector}

\item{k}{number of historical data available; integer}

\item{tau}{reports if a false alarm occurs before tau observations; integer}
}
\value{
TRUE if a false alarm occurred
}
\description{
Internal functiion that generates data and reports if a false alarm was made
by the Sequential Jump and Kink Detector before tau observations.
}
\keyword{internal}
