% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{initialize_Tstat}
\alias{initialize_Tstat}
\title{Intialize test statistic}
\usage{
initialize_Tstat(N, change_type, hist_data, pred)
}
\arguments{
\item{N}{window size}

\item{change_type}{"jump" or "kink"}

\item{hist_data}{vector of historical data}

\item{pred}{prediction of the linear function on the historical data}
}
\value{
the test statistic
}
\description{
Internal function to initialize the test statistic
}
\keyword{internal}
