% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdtuning.R
\name{M_quantile}
\alias{M_quantile}
\title{Maximal test statistic quantile}
\usage{
M_quantile(M, t)
}
\arguments{
\item{M}{matrix of maximal test statistics}

\item{t}{possible threshold}
}
\value{
the quantile that is reached with this threshold
}
\description{
Internal function returns the quantile of the threshold and
maximal test statistics
}
\keyword{internal}
