% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdtuning.R
\name{max_Tstat}
\alias{max_Tstat}
\title{maximal test statistic}
\usage{
max_Tstat(N, change_type, tau, k)
}
\arguments{
\item{N}{vector of integers; window sizes}

\item{change_type}{vector of "jump" and "kink"; type of algorithm}

\item{tau}{integer; number of test statistics to calculate}

\item{k}{integer; amount of historical data}
}
\value{
maximal test statistic
}
\description{
Internal function that calculates the maximal test statistic
}
\keyword{internal}
