% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdtuning.R
\name{max_Tstat_N}
\alias{max_Tstat_N}
\title{maximal test statistic for single window}
\usage{
max_Tstat_N(N, change_type, data, tau)
}
\arguments{
\item{N}{window size; integer}

\item{change_type}{type of algorithm; "jump" or "kink"}

\item{data}{vector of data}

\item{tau}{number of test statistics to calculate; integer}
}
\value{
maximal test statistic
}
\description{
calculates the maximal test statistic for a single window
}
\keyword{internal}
