% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_detection}
\alias{plot_detection}
\title{Plot detection}
\usage{
plot_detection(data, tau_0, alpha_minus, beta_minus, alpha_plus, beta_plus)
}
\arguments{
\item{data}{vector of data points}

\item{tau_0}{true changepoint; integer}

\item{alpha_minus}{intercept at tau_0 of the first segment; numeric}

\item{beta_minus}{slope on the first segment; numeric}

\item{alpha_plus}{intercept at tau_0 of the second segment; numeric}

\item{beta_plus}{slope on the secodn segment; numeric}
}
\value{
plot
}
\description{
plots data, the signal and the true changepoint,
endpoint is the time of detection
}
