% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{seg_lin_fun}
\alias{seg_lin_fun}
\title{segmented linear function}
\usage{
seg_lin_fun(x, tau, alpha_minus, beta_minus, alpha_plus, beta_plus)
}
\arguments{
\item{x}{points to evaluate the function at; vector of integers}

\item{tau}{changepoint; integer}

\item{alpha_minus}{intercept at tau of first segment; numeric}

\item{beta_minus}{slope on first segment; numeric}

\item{alpha_plus}{intercept at tau of second segment; numeric}

\item{beta_plus}{slope on second segment; numeric}
}
\value{
vector that describes a segmented linear function
}
\description{
A function that returns a segmented linear function at times x
}
