% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{SeqJaKD_initialize}
\alias{SeqJaKD_initialize}
\title{Sequential Jump and Kink Detection Initialization}
\usage{
SeqJaKD_initialize(hist_data, N, change_type, t)
}
\arguments{
\item{hist_data}{vector of historical data for the initialization}

\item{N}{vector of window sizes}

\item{change_type}{vector of strings "jump" or "kink" defines the type of algorithm
for the corresponding window size}

\item{t}{numerical vector of thresholds for the corresponding window sizes}
}
\value{
list with
detection - always set as False
data - The data window needed in storage for the first iteration
Tstat - the test statistic
N - vector of integers that describes the window sizes
change_type - vector of strings "jump" or "kink", defines which algorithm is
used for the corresponding window size
t - numerical vector that describes the thresholds to the corresponding
window sizes
slope_est - estimate of the slope of the pre change mean from the
historical data
}
\description{
Initializes the Sequential Jump and Kink Detection for SeqJaKD_iterate
}
