% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{SeqJaKD_iterate}
\alias{SeqJaKD_iterate}
\title{Sequential Jump and Kink Detection Iteration}
\usage{
SeqJaKD_iterate(new_data, detector)
}
\arguments{
\item{new_data}{the new data point}

\item{detector}{list provided by the last iteration or the
SeqJaKD_initialize function}
}
\value{
list with
detection - True if a detection was made in this iteration
data - The data window needed in storage for the next iteration
Tstat - the test statistic
N - vector of integers that describes the window sizes
change_type - vector of strings "jump" or "kink", defines which algorithm is
used for the corresponding window size
t - numerical vector that describes the thresholds to the corresponding
window sizes
slope_est - estimate of the slope from the pre change mean
}
\description{
An Iteration of the Sequential Jump Detection and the Sequential Kink
Detection Algorithm
}
