% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdtuning.R
\name{t_tuning}
\alias{t_tuning}
\title{threshold tuning}
\usage{
t_tuning(N, change_type, tau, eta, k, r)
}
\arguments{
\item{N}{vector of window sizes}

\item{change_type}{vector of "jump" and "kink" deciding the algorithm used for
each window size}

\item{tau}{integer that defines to which time the algorithm has to run with
a probability of eta without a false detection}

\item{eta}{numerical in (0,1) that is the probability with which the algorithm
runs until eta without a false detection}

\item{k}{integer that decides the number of available historical data}

\item{r}{integer that decides the number of maximal test statistic that is
drawn to decide the threshold}
}
\value{
vector of tuned thresholds
}
\description{
Tunes the threshold for given window sizes to bound the type I error such
that the probability of a false alarm before tau is approximately eta.
}
