% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{update_Tstat}
\alias{update_Tstat}
\title{Update Test statistic}
\usage{
update_Tstat(new_data, old_data, old_Tstat, N, change_type, slope_est)
}
\arguments{
\item{new_data}{new data point}

\item{old_data}{window of the last data points}

\item{N}{window size}

\item{change_type}{"jump" or "kink", defines the algorithm used}

\item{slope_est}{estimate of the slope}
}
\value{
the updatet test statistic
}
\description{
Internal function that updates the test statistic in SeqJaKD_iteration
}
\keyword{internal}
